Added login form and fixed css style
Added login form ONLY for worker. Also the form doesn't work as it's
supposed to, but it does check if the combination of workerID and
password are correct. If they are, you are redirected to another page,
but if they're not, then the worker has to login again. (Disclaimer: I
have not added any messages to notify worker whether their id and
password combination are valid.)
Also added css files and did some minor changes in header.